Add login link to layout
The application layout file was copied to the videohoster module and
a link to the login route was added in the navigation menu. The
application layout was overridden as I don't want to pollute the
module. I want to keep as much of the work in the videohoster module
as possible.