Add the manage video form along with basic tests
This commit adds the basics of the managevideoform to the application,
along with basic tests for it. Currently the form's elements are
initialised in the class constructor, because the test fails if I
initialise them in the init method. I need to do some research to find
out what's going on there. Actually, it's likely because the init
method's called by the FormElementManager, which this test isn't doing.